CPU warping support.
